backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==3.2.2
cffi==1.16.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.4
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.18.1
python-dotenv==1.0.1